Chore: 알고리즘 항목 선택 로직  커스텀 훅으로 분리
