+++ src/pages/PostsPage/index.tsx
@@ -5,8 +5,8 @@ import BottomNav from '@/shared/layout/BottomNav';
 import PageHeader from '@/shared/layout/PageHeader';
 import FloatingButton from '@/shared/ui/FloatingButton';
 import SearchInput from '@/shared/ui/SearchInput';
-import { useState } from 'react';
 import AlgorithmDropdown from './ui/AlgorithmDropdown';
+import useAlgorithmDropdown from '@/shared/hooks/useAlgorithmDropdown';
 interface IAuthor {
   userId: number;
   nickname: string;
@@ -59,24 +59,13 @@ const hotPost = {
 
 const PostsPage = () => {
   const { onChange, value: searchValue, reset: resetInputValue } = useInput();
-  const [selectedAlgorithmTypes, setSelectedAlgorithmTypes] = useState<string[]>([]);
+  const { selectedAlgorithmTypes, handleToggleAlgorithmType, handleClearAllTypes } =
+    useAlgorithmDropdown();
 
   // 검색 처리 & input 값 초기화
   const handleSearch = () => {
     console.log('검색한 값', searchValue);
     resetInputValue();
-  };
-
-  // 알고리즘 유형 모두 선택 해제
-  const handleClearAllTypes = () => {
-    setSelectedAlgorithmTypes([]);
-  };
-
-  // 알고리즘 유형 선택/해제
-  const handleToggleAlgorithmType = (type: string) => {
-    setSelectedAlgorithmTypes(prev =>
-      prev.includes(type) ? prev.filter(t => t !== type) : [...prev, type]
-    );
   };
 
   return (
